Developed the bitmap interaction software
Using various streams and file handling mechanisms, this addition to my NEA pulls an image into memory from a specified file location, converts the image to a bitmap, splits it into a header and the actual pixel data(which can be interacted with directly) and after some editing will rejoin the two components and save them as a seperate bitmap file. To test the software I have inverted the colour of the test maze, made it a red version of itself, and made it completely red.