New methods + explanation.txt
Created a txt file with some explanations about what the main variables mean and how I have done things (Intend to add more to it). Implemented a method to grayscale a bitmap using the method on wikipedia. Implemented a method to make every pixel either black or white depending on which side the grayscale value is on the _bwthreshhold value(Intend to do testing to work out a more universally functional value might be or perhps think of an algorithm to determine the value based on the image given).